feat(userzone): 个人中心信息解析

+++ src/main/java/top/xcphoenix/groupblog/model/dao/Category.java
@@ -12,7 +12,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class Category {
 
-    private long categoryId;
+    private int categoryId;
     private String name;
 
 }
